Shrink holes on all switch footprints
The holes were generously large, to accommodate any possible margin of
error. This caused a small amount of "wiggle" when testing the switches
on sample PCBs.

Since the dimensions were based on datasheet footprintrecommendations,
it makes sense to simply reduce them to those exact values. For the LED,
the updated dimensions were based on generic footprints for
similar-sized LEDs.