Add symbol and footprint for TCS3472
Footprint is similar to the DFN-6 used by the TSL2591, but the datasheet
recommended some very slight pad size differences. Also attempted to use
the square/oval pad shape shown in the datasheet.

The 3D model is a placeholder, which fits the footprint. It does not
accurately reflect the specific nuances of the sensor design itself.