Implemented a dedicated symbol/footprint for the NHD-3.12 display
The footprint combines the pin socket/header connector with correctly
aligned mounting holes, to make PCB layout easier.

The symbol labels the pins for the display module's serial interface, to
make the schematic a little bit nicer.

Included the STEP model of the display from the NHD website.

Functionally, its not that different than before.