Rotary encoder input implemented
Detecting rotation with the timer encoder feature on TIM1
Only triggering on even counts, since we get two interrupts per tick of
the knob.
Still some debouncing(ish) issues to resolve, where we'll get a lone event in
the wrong direction when fiddling with the knob. Likely a quirk of the
input we can filter out, but something that should be dealt with.

Detected rotation ticks are injected as key events into the keypad event
queue.

No idea what the TIM1 commutation interrupt is used for, but had to
disable it since it occasionally fires and is at a priority level
incompatible with FreeRTOS interrupt handling.