Redesigned keypad handling to support key combos and repeat events
Originally, the keypad handler would do its thing directly on the GPIO
event task and send results to the key event queue. The problem was that
this couldn't easily deal with timers or other behaviors.

The new design adds a dedicated keypad task, which acts as an
intermediary. Raw events from the GPIO event task function (which runs
in response to controller interrupts) are added to an internal queue.
This queue is then handled by the dedicated keypad task to produce
events handled by the rest of the application.

Additionally, this task/queue allows the addition of keypad timer
events, which are used to implement key repeat behaviors.

On top of this, some initial changes to the basic keypad handling in the
main task has been done. These will be followed up and improved later.