Initial refactoring to a state controller
Created a controller module to handle the overall state machine of the
device, starting with basic switch-case style state transition
management.

Moved existing functionality into the new structure.

The eventual plan may be to use a more robustly designed state machine
manager, and this implementation will be coded out with that in mind.
However, its easier to start this way to figure out what each state will
need to deal with.