Fixed encoder interrupt handling to be safer
Somehow the commutation interrupt was getting triggered even though I
thought it wasn't being enabled. This breaks FreeRTOS due to its
priority level. So instead, the new approach is to let it always be
enabled, but make sure we're not actually handling it in the code that
dispatches events into FreeRTOS.