Initial implementation of exposure countdown timer
This implementation uses a simple while-loop with tick counter checks to
manage everything, and has simple display drawing optimizations. It
doesn't toggle the relays yet either.

The goal here is simply to rough out the user-visible behavior.

The ultimate goal is to switch over to using system timers for
synchronizing the countdown clock, relay toggling, and beeps. Doing this
will enforce more accurate and repeatable times for all of the above.