Basic test strip mode implementation
A basic implementation of the UI parts of test strip mode, for
incrementally-covered test strips. This is to figure out the display
elements and user interaction.

This code will need significant refactoring and possible design changes
for cleanup once a better approach is figured out for the main exposure
timer.