Exposure timer rewrite to use interrupt-based time scheduling
Implemented a new mechanism for handling the exposure timer, which uses
interrupts to keep timing consistent and reliable. It works by
configuring TIM10 to fire on a 10ms interval, then using the associated
ISR to control the events of the exposure process.

These events include turning the relay on and off, poking the
user-visible countdown clock, and generating tick beeps every second.

The user notifications happen by having the task block on notifications
generated by the ISR, and having the result of those notifications cause
a callback function to be invoked.

This new timer is implemented in a standalone module, and now integrated
into the main exposure and test strip modes.