Add focus mode with timeout
No longer making this its own state, since it makes more sense as a
feature of the "home" state that can remain on through other
home-associated states.

Also implemented a timeout, which is reset on key events, to make sure
the enlarger does not remain on indefinitely when nothing is happening.
This timeout will eventually be user-configurable.