Initial rough implementation of enlarger time profiling
Created an automated process to profile the behavior of the enlarger
lamp, in relation to activation and deactivation of the control relay.
Right now this process is designed to primarily produce experimental
results that can be recorded manually, but it will eventually be
incorporated into a process for building enlarger profile
configurations.

The process begins by measuring light levels that can be considered
"enlarger on" and "enlarger off" to get reference points. These levels
are sampled many times, and then tracked as a set of statistics.

The next step is to run the enlarger through a series of on/off cycles,
tracking exactly how long the following phases take:
- Relay activation to light level rising
- Light level rising to full brightness
- Relay deactivation to light level falling
- Light level falling to fully dark

With all the above data, the "equivalent full exposure" time is then
calculated for the rising and falling periods. All of this data is then
displayed to the user.

Additionally, a data structure for containing this data has also been
implemented and incorporated into the exposure timer process. However,
the link between these two parts does not yet exist and chosen profile
values currently need to be hard-coded in the app settings code.

The goal of this process is to make it so that the user-visible
"exposure time" represents an effective exposure at full brightness,
rather than the actual time between relay on and relay off. This
distinction can become important with short exposures and incremental
test strips, where the behavior of the lamp can make the actual exposure
time effectively 1/4 to 1/2 of a second shorter than the relay on/off
time.