Initial blackout mode implementation
Blackout mode is based on the position of the blackout rocker switch on
the front of the device. If the mode is enabled, then the safelight
relay and all illumination from the device (display and LEDs) is turned
off. None of the other device functions are inhibited, however.

This mode is intended to be used for temporary periods while handling
materials that are sensitive to all wavelengths of light, such as color
paper.

Currently, there may be a brief LED flash on startup even if blackout
mode is enabled, because some parts of device initialization happen
before the switch state is known. Also, it is possible to mess up the
blackout state via certain diagnostic functions. This should either be
prevented or restricted during blackout mode.

Given that it is likely to be quite easy to get lost in some
ancillary device function while in blackout mode, the allowable
states should probably be limited.

Additionally, it may help to provide extra audio cues when in this mode.

As implemented, the blackout toggle is independent of all other user
functions, and should be freely selectable at runtime.