Increase reliability of encoder rotation events
This is an attempt at filtering out the wonky results that were coming
from a more simplistic handling of encoder counter interrupts.

To generate an event, we now must receive two interrupts that both show
the encoder rotating in the same direction. Furthermore, these
interrupts must happen within a brief period of each other.

Doing this appears to keep encoder events to be consistent with
intentional detent "clicks" by the user, and seems to prevent spurious
counter-direction events.