Implement burn/dodge exposure timing
This adds the burn/dodge exposure timing process onto the end of the
main exposure sequence. The beeping and timing is similar, except for
requiring additional presses of the Start button. (This may change in
the future).

The UI for this consists of an interstitial screen showing the
adjustment index, value, and associated contrast grade. This screen then
transitions to a similiar screen with a clock icon to show the actual
countdown.

When doing a dodge, the base exposure time is reduced.