Refactor 7-seg digit drawing to reduce code duplication
Each digit is composed of a collection of segments, and the identifiers
for those segments are identical regardless of digit size. It is
wasteful and redundant to have a separate set of switch/case statements
to identify segments and call draw functions for each digit size. As
such, the new approach breaks digit drawing into a few distict steps
with minimal duplication:

1. Outer wrapper functions providing the same interface as before.
2. Common draw digit function, that uses a switch/case to assign segment
   bit fields and call a provided segment draw function pointer.
3. Segment draw functions for each digit size, that test each bit flag
   and draw the segments that are enabled.