Reduce enlarger calibration value separation requirements
The defaults were actually too large to work for my enlarger at a normal
height, so this relaxes things a bit.

Ideally there should be some user indication of these numbers at the
start of the calibration process, to make it easier to adjust the
enlarger for optimal calibration (open the lens wide, lower the height a
lot, etc.) For now, this requires going to the meter probe diagnostics
to see what the readings are.