More correctly implement USB keyboard support
Change USB keyboard event routing so that it is no longer intermingled
with regular keypad events, even if its distributed via the same queue.
This now requires key handler code to separately handle events from both
sources, but has the benefit of allowing them to be treated separately.
It is necessary to do this to avoid confusing state between the
different input devices.

Update u8g2 display input handling routines to deal wih the USB keyboard
changes.

Modify display_input_text to support entry via the keyboard, and changed
the behavior of the "+" and "Menu" buttons on this screen for
consistency.

Handle capslock correctly in the USB keyboard code, with limited
handling of NumLock and ScrollLock.