Add ability to configure sensor GA factor
This makes meter probe sensor calibration configurable on-device, so a
hard-coded constant is not necessary. In the future versions of the
meter probe, it may make sense to store these values inside the meter
probe itself. For now, however, they're stored in the Printalyzer's
settings page.

Added a menu function to allow choosing a number with a decimal point.
Added a config option for the TCS3472 GA factor
Updated the diagnostics for the meter probe to support using, or not
using, the configured GA factor.