Menu for basic editing of paper profiles
Right now the profile edit menu is fairly simple, and mostly focuses on
giving the user a way to enter the raw numbers. This is all that is
technically required, but it is not the goal.

The goal is to supplement this with something that can calculate these
numbers from a set of measurements, perhaps from a step wedge exposure.
Doing that right will require a lot more data entry capabilities, so it
is being pushed off for the moment.