Add calibration mode to the device home screen
Pressing the left/right arrow buttons together allows for
mode-switching now, with a new "Calibration" mode being added alongside
the existing "Printing" mode.

In calibration mode, burn/dodge adjustments are disabled, and PEV values
are prominently displayed based on the last meter reading.

Taking a meter reading will calculate an exposure time to achieve a
decent starting PEV. From this point, exposure time adjustments will
update the displayed PEV accordingly. Also, the contrast arrows will now
adjust the PEV and update the exposure time as well.

When making a test strip, the display will show the PEV for each patch
of the strip.

The goal of all this is to make it easy to expose for a specific PEV, or
to find the PEV associated with a specific exposure or strip patch.
These values can then be used directly for building a paper profile.