Add max net density to paper profiles
Calculating the complete characteristic curve for a paper grade requires
knowing the density values corresponding to all three PEV points. Two of
these are fixed in the standard. The third is based on the paper's max
net density (relative Dmax).

This adds the value to the paper profiles, the profile editor menus, and
fixes a title bug with one part of the profile editor.