Initial implementation of tone graph and print metering
Tone graph is computed via one of two methods, depending on how much
information is available in the paper profile.
- By evenly dividing the paper's contrast range
- By evenly dividing the paper's density range, and using an
  interpolation of the paper's characteristic curve to find where those
  points fall on the contrast range scale.

Metering sets the base exposure to whatever places the lowest reading at
the D=0.04 (lower bound) point, then populates the tone graph based on
this and the rest of the values.

There are a few outstanding issues with this process that should be
fixed in future commits:
- Sometimes tones jump between separate and overlapping scale patches,
  likely due to floating-point comparison issues. This is expected when
  changing contrast grades, but should be avoided when simply changing
  exposure time.
- There should be a brief visual "blinky" effect showing the most recently
  added tone on the scale.
- Identical readings should be merged, and not count separately (within
  some reasonable tolerance.)
- The tone graph should remain visible on some other screens, like fine
  adjustment.