Implement basic support for FTDI USB serial adapters
This is a fairly basic driver implementation, as our actual needs with
these things is rather simplistic. It may not hold up under stressed or
varied use cases, but it does work.

This adds support for using a variety of FTDI USB serial adapters, using
the ChibiOS and FreeBSD/NetBSD driver implementations as a guide and
source of relevant protocol constants.

The baud rate and chipset type code should probably be rewritten at some
point to more closely match the FreeBSD/NetBSD version, as it currently
matches the simpler ChibiOS version.

This code has also only been tested on the FT232R chipset, but should
theoretically work on many more FTDI variants.

When additional USB serial adapters are added, modifications to the STM
USB host library will be necessary as they're all likely to be
vendor-class devices.