Implement basic support for PLCOM USB serial adapters
This is a fairly basic implementation, mostly tested against a modern
PL2303. It doesn't cover the enormous swath of potential vendor/device
IDs that may exist in the real world. It is designed for our use case of
sending/receiving limited bits of text, so it may not hold up under
stressed or varied use cases.

It is primarily based on the initialization and protocol bits from the
FreeBSD/NetBSD driver for the same chipset, though it uses constants and
structures from ST's CDC class header where relevant.

In addition to the driver itself, a number of changes were necessary to
supporting code:

The ST USB Host Library's class implementation selection function was
modified to support multiple vendor-class devices being registered at
the same time. It tries them in sequence, silently moving on to the next
one if the current one returns a USBH_NOT_SUPPORTED error from its init
function.

The ST HAL's low level USB support was modified to add a configuration
flag to its channel settings. This flag changes the behavior of its
interrupt handler.  By default, in response to a CHH/NAK interrupt, it
restarts the channel prior to returning the state change to the
application. On the PL2303, this restart causes the interrupt to fire
in a tight loop until data is received. The added flag disables this
auto-restart behavior, making it possible to safely poll the USB device
without causing the whole system to grind to a halt servicing
unnecessary interrupts.