Implement basic support for SLCOM USB serial adapters
Like the FTDI and PLCOM drivers, this is also a basic implementation
designed for the simple use cases of this project.

This chip is far simpler to initialize than those other chips. The
FreeBSD driver was used as a reference for this initialization process,
and for some of the constants/commands, but a lot of the code was
specific to this driver.

Unlike the other chips, this one also has usable official documentation
from which a few other features (not in the FreeBSD driver) were added.
Most important of these is the ability to query for received characters
prior to starting a bulk read. This query avoids the need for using the
HAL "workaround" that was required for the PLCOM driver.

The documentation and the FreeBSD implementation also discusses a
control pipe that's used for bits of status information. Use of this was
mostly punted on, since its not necessary for our simple use cases.