Building out USB Serial public interface
Implemented a standard config interface for line coding settings

Implemented blocking transmit that behaves as expected

Implemented a line-terminated receive function that uses an internal
buffer and can wait for a specified amount of time. It is expected that
this function will be polled with short blocking periods alongside other
functions such as the keypad queue polling routine.