Basic implementation of reading densitometer messages
This is a stateless implementation of polling the USB serial receive
buffer, and parsing out densitometer readings. Data formats for both
Heiland and X-Rite units are supported, and auto-detected as well as
possible.

Also added a diagnostic screen to show the results of this interface for
setup testing purposes.

It may make sense to also add serial port configuration to the system at
some point, but for now a default 9600N1 config seems workable as a
universal config.

While a fancier bi-directional interface could be implemented in some
cases, this version should be sufficient for the initial use cases.

Resolves #6