Add ability to feed a densitometer reading into Dnet
Make the "Max Net Density" paper profile setting able to change its
value directly from a densitometer reading.  This is only really
appropriate for densitometers that can be "zero'ed" on the paper base,
but is the only existing area in the UI where we can use this
capability.

It is mostly a proof of concept and a lot of plumming work to allow
for feeding densitometer readings into numeric entry display screens.

Fully building out the ability to create paper profiles with
densitometer measurements will be a dedicated set of screens a lot of
unrelated UI work.