Add step wedge configuration and settings UI
This adds the necessary structures, settings code, and UI code, to
handle a step wedge profile. Only a single configured step wedge is
supported, but that should be sufficient for the expected use cases.

The intended use of the step wedge profile is paper calibration, where
it will be applied to data from an as-yet-unbuilt step wedge exposure
module under the paper profile settings to automatically calculate paper
profile values.

The step wedge profile serves no functional purpose outside of the
process of creating a new paper profile, but it has to go somewhere
sensible in the menu system. Its currently a top-level item to prevent
it from being hidden where no one can find it.