Implemented paper profile calculations from step wedge exposure
Added a screen for user input of densitometer measurements of paper from
a step wedge exposure. Individual input values can be captured directly
from a serial-attached densitometer, but its still a one-at-a-time
process.

Implemented the calculation of a paper's characteristic curve from these
measurements. The process goes through the following steps:
- Capture the paper base density (if the densitometer cannot be zero'd
  on the paper base), the step wedge properties, the print exposure
  value, and patch measurements from the above screen.
- Calculate the equivalent PEV values for each test exposure patch, by
  applying step wedge data to the base PEV.
- Use the CMSIS-DSP library to perform cubic spline interpolation across
  the data, which yields the full characteristic curve of the paper.
- Iterate across this interpolated graph to find PEV values that
  correspond to the 3 density points (Ht, Hm, Hs) used for our paper
  profiles.

Implemented a results display screen that shows a graph of the above
curve, scaled for display pixels, next to a preview of the resulting
numbers.

If selected by the user, these numbers are copied back to the paper
profile.

Still need some way of easily changing the default PEV target for the
Printalyzer's calibration mode, because step wedge tests work better
with a PEV in the ballpark of 200. The current default of 30 is only
appropriate for finding the Ht value independently.