Add initial build tagging to firmware
This puts in place some initial bits to include the build date and git
information inside the firmware image. It is just a starting point, and
may need refinement to get user-acceptable version numbers in a
repository that may have hardware and software revision tags mixed
together.

While the macros added to the build are setup in the makefile.defs file,
they needed to be added to the compiler directly inside the Eclipse
build configuration.  Trying to put them all into a single variable for
inclusion on the Eclipse side had the unintended effect of breaking
Eclipse's system include file indexer.

Fixes #8