Initial integration of bootloader support into the project
Added in a basic CubeMX code dump, with configuration changes to support
bootloader needs. This means no FreeRTOS, USB host support just for MSC,
and enabling the CRC module. Also, the HAL timebase has been changed
back to SysTick and unnecessary timer peripherals are no longer
configured.

Pulled in a cut down version of the u8g2 library. This means that
instead of dropping the code in as-is, just the csrc portion has been
added as a pair of "src" and "include" directories. Furthermore, display
driver code for all display drivers besides the NHD/SSD1322 has been
removed. Also, all fonts except a version of "pressstart2p" have been
deleted. This was done to avoid repository bloat, and to keep this
maintainable as a completely-independent "fork" of the main firmware's
libraries.

Updated linker and stack pointer configurations to give the bootloader
128K of space and the remaining 384K to the main firmware. This split
may be changed in the future, once the bootloader is complete and can be
compiled with space optimization, but its an acceptable starting point
for now. The way flash sectors are arranged on the STM32F4, making
changes here will require updating numerous bits of code across the
project.

Added in basic init/handler code for USB, keypad, and display. These
versions will only have the absolute most minimal features needed for
our purposes and not use anything that requires FreeRTOS components.
They will be expanded as needed.

Adding in bootloader files originating from Akos Pasztor's STM32L4
bootloader project. Since this bootloader implementation has required
extensive porting due to differences between the STM32L4 and the
STM32F4, a lot of the flash handling code had to be somewhat rewritten.
In the process of porting, all the comments and naming conventions were
changed such that the code no longer directly resembles the original
source.

Added checksum generation to the main firmware build process. Doing this
was somewhat complicated, since it required using a partially
STM32-specific CRC-32 algorithm, running it on code that was padded out
in the exact same way as the ST-Link flash loader would pad things (0xFF
padding instead of 0x00 padding). Integrating the checksum also had to
be done by making it an extra section of the .elf linker output file,
otherwise the STM32CubeIDE tooling wouldn't be able to load it properly.

Finished out a basic initial implementation of firmware updates from
USB. If a file with the right magic name is found, is of the right size,
and has a valid checksum, it will be programmed into the flash. There is
also some limited user information displayed during this process.

Improving the quality and functionality of the bootloader UI will be
left up to a future task.

Fixes #9