Upgrade firmware code (main project) to V1.26.0
This pulls in all the changes from upstream, minus any diffs caused by
relevant local edits. The USB host middleware code isn't included,
because that will be merged in a future commit. (Upstream USB is ahead
of the STM32F4 FW package.)

The easiest and safest way to handle these sort of merges right now is
to use the "import-cube.sh" script to dump a CubeMX export into a fresh
directory with the same structure as the project, and to then use a
directory-level diff tool to review and merge all the changes.

Refs #24