Pulled in changes from USB Middleware v3.3.5
The majority of the changes were actually things already done inside
this codebase to silence compiler warnings, however there were a couple
minor ones on top of that.

This was a more manual process than the FW package update, simply
because this isn't covered by the script. Hopefully this differing from
the FW package will generally be more of an exception than the rule.

Refs #24