Basic implementation of device settings import/export
Provides the basic ability to dump all device settings and profiles to a
JSON-formatted file on an inserted USB stick.

Also provides the ability to import these settings from the USB stick,
selectively deciding which portions to import. This may help with making
it possible to work with paper profiles off-device, without overwriting
other device settings.

Right now this process uses a fixed filename for simplicity, but that
should be changed as soon as a file browser can be implemented.

Fixes #1