Refactor and change title rendering of value input functions
Refactored the common elements of all the display value input functions
to a common implementation. It doesn't look any prettier, but cleans up
a lot of code duplication.

Changed the display value input functions to separate the title and
message text portions similar to how selection lists work. This makes
the title for those screens a lot clearer, and removes the need to do a
lot of inline formatting to make titles more obvious.

Also removed FIXME comments for a chance that's probably not going to
get made. If the increments for F16 input are increased in the future,
that should probably be another parameter since its usefulness is
context dependent.