Added utarray header from the uthash project
Modified utarray to use FreeRTOS functions for memory allocation

Rewrote out-of-memory (OOM) handling in utarray.

This is a simple change to how utarray deals with malloc failures.
Instead of calling a function we can't handle in any way other than as a
fatal error, simply allow the current operation to safely end and make
it possible to check the state of the result to find out if an OOM
happened.

This isn't quite as good as whatever fancy method uthash uses now, but
its hopefully good enough for our purposes.