Implemented a filepicker component
This is a general purpose filepicker to be used anywhere we need the
user to select a file from a USB storage device.

There are directory size limits that cannot be dealt with until the
display code is rewritten to support >8-bit selection lists, but this
may be an acceptable limitation for the use case.

For convenience of memory allocation, paths are returned in the 8.3
"altname" format. Since this isn't ideal for user display, a utility
function to re-find and get the LFN name for a file has also been
provided.