Add cancel confirmation to the settings menus
All settings menus with a multi-step (easy to accidentally cancel) flow
now have confirmation prompts if the user tries to leave them without
saving changes.

To help with this, a utility function was added to facilitate the
prompt.

Also fixed a deadlock issue that crept into the display brightness menu.
This one may be a case where our display mutex should really be
recursive, but I don't think there are other examples of this as of yet.

Refs #22