Add ability to runtime-switch paper profiles
This adds a long-press action to the "test strip" button that allows
changing of the active paper profile without any persistence. It may not
be the ideal placement for this action, but we're out of other good
options (at least as long as we don't want the menu system to require a
long-press to enter).

Fixes #3