Implement a flexible app descriptor to replace the image checksum
This implements a new application descriptor structure that contains a
variety of relevant information about the application build, in addition
ot the simple checksum that existed previously. It is fixed at a known
location within the binary image, so that it is readable by both the
regular firmware that contains it and the bootloader. The checksum
address itself is unchanged.

The new approach to doing this removes a fair bit of build-time hacks
that were done in favor of finally figuring out how to make the GNU ld
linker script behave correctly to our needs. The changes to that linker
script now do the following:
- Make sure padding between the ISR vector and the start of the
  application code in the ELF file actually match what the STM32 flash
  loader puts there (0xFF instead of 0x00).
- Pad out the end of the image with the same 0xFF data.
- Place the app descriptor structure in the last 256 bytes of the image,
  done at linking time rather than as a post-build kludge.

The "checksum.pl" script is now updated to work a little differently.
Instead of just calculating and writing out a checksum of the image, it
now writes out an updated version of the entire app_descriptor section
of the binary. It also no longer needs to fake any padding bytes, due to
the linker script changes above.

The custom Makefile targets have now been simplified. They go through
the same basic set of steps, but the implementation of those steps is a
lot simpler and no longer generates warnings. The process basically
involves:
- Run the "checksum.pl" script on the BIN file to generate an updated
  app descriptor section which contains the CRC-32 checksum
- Run "objcopy" to make a new version of the ELF file which contains
  this updated descriptor section.
- Run "objcopy" again to make a new version of the BIN file based on
  this updated ELF file.

In additon to all of this, a few minor changes were made to the macros
used for embedding certain build properties into the descriptor.
All relevant bits of application code that depend on descriptor
information have also been modified to get it straight from the
descriptor structure now.

None of these changes break the current bootloader, but the bootloader
should be updated to take advantage of them when possible.

Refs #23