Update bootloader to use the app descriptor structure
This is a minor update to the bootloader to use the new app descriptor
structure when checking for the CRC of the loaded firmware, and to read
that structure out of a firmware file to improve its info display before
flashing it.

There is no real functional change to the bootloader yet, but this is a
good first step.

Refs #23