Refactor display element data structures to clean up code
The three different modes of the home display now have separate display
functions and seperate argument data structres (and utility functions).

This reduces the amount of needless if-then logic scattered across the
exposure state to display path.

It might also make sense to break this into three separate state
controller states, to clean up the key handling logic, but that'll be
left as a possible future task.