Show an active tone graph on the burn/dodge adjustment editor
This change makes it easier to visualize how a burn/dodge adjustment
will affect the measured tones of the image. The basic visualization is
similar to what's being done for the fine time/stops adjustment states.

When the burn/dodge is being done at a different grade than the base
exposure, the graph is recalculated accordingly. Handling this did
require a fair bit of refactoring to allow using a temporary version of
the graph marks array. As currently implemented, there is thus a bit of
a performance penalty when displaying these working-copy modified tone
graphs. However, this may be small enough to not seriously impact the
UI. If it becomes an issue, this code path can be optimized.

Fixes #29