Added indication of main exposure time being too short
Based on the active enlarger profile, it is possible for the exposure
time to be too short to be timed reliably. In these cases, we should
alert the user to this.

The first step was to fix some bugs in how this state was made known to
the state controller and exposure state. Somehow it was not being set
correctly when loading the active profile, and wasn't properly being
converted from milliseconds to seconds. That issue should now be fixed.

Then an indicator was added to the main printing and calibration timer
displays when the displayed exposure time is less than what the enlarger
profile can reliably support.

There still needs to be indicators for burn/dodge use cases and some way
of handling the test strip case. These are the most common situations
where this issue will arise in the real world.

Refs #25