Add short exposure time indicators for burn/dodge
This adds the indiciator to the intermediary screen shown during the
main exposure sequence, and to the detailed burn/dodge editor.

For burning, its a simple calculation based on the burn time. For
dodging, the indicator will appear if either the dodge time is too
short, or the reduced base exposure is too short.

Refs #25