Fix meaningless ERC issues
KiCad 6.0.10 made some power library changes that led to ERC warnings
all across every existing schematic.  This should fix the issues for
this project.

There's still one GND net related warning on the main board, which will
be resolved as part of a future task to review the grounding strategy
for that board.