Initial split of main board into separate main and power boards
This is a first pass at splitting the main-board project into two
separate projects. There's still a lot of placeholder bits around the
inter-board connection, and PCB routing is incomplete.

The main-board now excludes anything dealing with mains AC power, or the
device's rear panel.  As a result, it got a bit shorter.

The new power-board contains all of the bits that deal with mains AC
power input, relay power switching, and anything else that might need to
connect to the rear panel of the enclosure. It is also now only a
2-layer board, and it will stay that way unless there is an explicit
need to do otherwise.

The purpose of the separate power board is to isolate anything that
deals with mains AC, so it can be reviewed/inspected/updated
independently. Having it separate will also make it possible to reduce
the size of the enclosure, especially depth-wise.