Update OLED display footprint to better match alignment needs
Footprint is now centered on the dead center of the display's visible
area.
3D models now assume a 3mm spacing between the bottom of the display
module and the PCB.
A 3D model representing the mounting spacers is now included. Its a very
simple shape that was custom created, so its not mentioned in the models
readme.

This display has 3mm mounting holes, but its probably safer to assume
M2.5 screws for mounting flexibility.  Mounting precision will be
provided by the header pins.