Change panel illumination LEDs
This changes from a single LED for each spot, to a combination of LED
and spacer.  The purpose of this change is to get a little bit less
height on the LEDs, while keeping the same overall specs, for better
enclosure fit.

The specific components are:
LED: SunLED XLUR11D
Spacer: Bivar ELM 2-120