Reduce main board width
Reduce the main board width from 115mm to 110mm, which should help
reduce the enclosure depth by at least 5mm, perhaps more.

Reroute relay control signals to go below the power regulators.
Move the DMX512 header downwards.
Nudge the 3.3V regulator downwards a tad and move the 12V test points.

This change does put the display mounting holes a fair bit closer to the
board edge. However, there are so many mounting points for that module
that this is unlikely to be cause for concern.