Add footprint for USB-A connector
This is a connector with SMD signal contacts and thru-hole shield
connections for mechanical strength. Chose the footprint with
oval-shaped holes for the shield and used a pin-in-paste approach for
the paste layer.

This should make the connector mostly soldered in via SMD reflow, at
most only needing minor touch-up for the shield.