Add enclosure design file for the RevB prototype
This file was created with Protocase Designer, and can be opened with
that tool.

The older FPD files in this directory are no longer valid for the
current design, and the models.md file's contents are now out of date.
These will both be cleaned up in a future commit.