Various fixes to get bootloader and app verification working again
Apparently the newer versions of GNU ld are now inserting a blank word
between the .text and .rodata sections, which doesn't get programmed as
expected. This causes a checksum failure.

The current fix is to use the "--gap-fill" command line option for
objcopy when generating the data to feed to the checksum generator, and
again when generating the actual output of the build process.  This
feels a bit kludgy, but works well enough to get things moving again for
now.