Fix SPI frequency and GPIO speed settings
A lot of these were maxed out, which was causing problems. Going over
the datasheet limits for the display was causing the display to
sporadically fail, and the GPIO rise times did not need to be anywhere
near as fast for anything else either.

Technically the datasheet claims a 100ns max clock rate, which I think
translates to 10MHz, but the closest we can actually select is 12MHz.
Going with that for now, which hopefully won't cause problems.