Strip out legacy I2C utility code
This was leftover from an older project where it needed to do more than
simply wrap a platform function. We might as well get rid of it now.

Temporarily copied the functions into the driver for the TCS3472,
because there's no point in updating that code as it'll soon be removed.