Initial DMX512 controller task implementation and diagnostics
This is the first pass at implementing a task to handle DMX512 frame
transmission, with a control interface. There is still plenty of room
for improvement of the interface, so this is just a starting point for
testing.

Also added a very rough initial diagnostic screen for controlling this
task, which simply allows for enable/disable and a few hard-coded value
patterns.